% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_supra_adjacency_matrix.R
\name{construct_supraadjacency_matrix}
\alias{construct_supraadjacency_matrix}
\title{Construct supra-adjacency matrix}
\usage{
construct_supraadjacency_matrix(weighted_layer_df, graph_list)
}
\arguments{
\item{weighted_layer_df}{A data frame containing weighted layer information.}

\item{graph_list}{A list of igraph objects representing network layers.}
}
\value{
A supra-adjacency matrix representing the connectivity between nodes
across multiple network layers.
}
\description{
This function constructs a supra-adjacency matrix representing the connectivity
between nodes across multiple network layers.
}
