% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_degree_dist}
\alias{plot_degree_dist}
\title{Plot Degree Distribution}
\usage{
plot_degree_dist(graph)
}
\arguments{
\item{graph}{The input graph object.}
}
\value{
None
}
\description{
This function plots the degree distribution of a graph as a scatterplot on a double logarithmic axis.
}
