% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_lcc_distribution}
\alias{plot_lcc_distribution}
\title{Plot LCC Distribution}
\usage{
plot_lcc_distribution(results_df, network_filter = NULL)
}
\arguments{
\item{results_df}{The dataframe containing results to be plotted.}

\item{network_filter}{An optional parameter to filter by network name. Default is NULL, which prints all histograms.}
}
\value{
None
}
\description{
This function applies and prints single histograms or all in the results dataframe. By default, it prints all histograms, but optionally, you can filter by network name.
}
