% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_lcc_overview}
\alias{plot_lcc_overview}
\title{Plot LCC Overview}
\usage{
plot_lcc_overview(results_df)
}
\arguments{
\item{results_df}{The dataframe containing results to be plotted.}
}
\value{
None
}
\description{
This function applies and prints an overview of all histograms for the input dataframe.
}
