% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{plot_z_score_ranking}
\alias{plot_z_score_ranking}
\title{Plot Z-Score Ranking}
\usage{
plot_z_score_ranking(results_df, alpha)
}
\arguments{
\item{results_df}{The dataframe containing results to be plotted.}

\item{alpha}{The significance level.}
}
\value{
None
}
\description{
This function plots the z-score ranking of observed LCC size to random expectation for each network as lollipop plot.
}
