% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{read_graphs_from_folder}
\alias{read_graphs_from_folder}
\title{Read Graphs from Folder}
\usage{
read_graphs_from_folder(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder containing TSV files.}
}
\value{
A list of igraph objects.
}
\description{
This function reads graph data from TSV files in a specified folder and
constructs igraph objects based on the graph type flag.
weighting is implicit when weight column is filled
}
\examples{
# Example usage:
# graph_list <- read_graphs_from_folder("path/to/folder")
}
