% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_multiplex_propagation.R
\name{weighted_multiplex_propagation}
\alias{weighted_multiplex_propagation}
\title{Weighted Multiplex Propagation}
\usage{
weighted_multiplex_propagation(
  supra_adjacency_matrix = NULL,
  graph_list,
  layer_weights_df,
  seed_set,
  seed_weights = NULL,
  restart_prop = 0.7,
  rank_results = FALSE,
  remove_seeds = FALSE
)
}
\arguments{
\item{graph_list}{List of igraph objects representing graphs for each layer.}

\item{layer_weights_df}{Data frame containing weights for each layer.}

\item{seed_set}{Vector of seed nodes.}

\item{seed_weights}{vector of seed, weights}

\item{rank_results}{Logical indicating whether to rank results (default: FALSE).}

\item{remove_seeds}{Logical indicating whether to remove seed nodes from results (default: FALSE).}
}
\value{
A dataframe containing the propagation results.
}
\description{
This function performs weighted multiplex propagation on a set of graphs
}
